+++ lib/src/models/query/includes.dart
@@ -1,0 +1,7 @@
+class Includes {
+  final List<String> includes;
+
+  const Includes(this.includes);
+
+  toJson() => includes.join(',');
+}